dir:Operator 관련 클래스 디렉토리 구조 변경
