+++ src/main/java/step2/operator/AddOperator.java
@@ -1,10 +1,10 @@
-package step2;
+package step2.operator;
 
 import java.math.BigDecimal;
 
-public final class MultiplyOperator implements Operator {
+public final class AddOperator implements Operator {
     @Override
     public final BigDecimal operate(final BigDecimal leftOperand, final BigDecimal rightOperand) {
-        return leftOperand.multiply(rightOperand);
+        return leftOperand.add(rightOperand);
     }
 }
