+++ src/main/java/step2/operator/DivideOperator.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.operator;
 
 import java.math.BigDecimal;
