+++ src/main/java/step2/operator/Operator.java
@@ -1,8 +1,8 @@
-package step2;
+package step2.operator;
 
 import java.math.BigDecimal;
 
 @FunctionalInterface
-public interface Operator{
+public interface Operator {
     BigDecimal operate(final BigDecimal leftOperand, final BigDecimal rightOperand);
 }
