+++ src/main/java/step2/operator/SubtractOperator.java
@@ -1,8 +1,8 @@
-package step2;
+package step2.operator;
 
 import java.math.BigDecimal;
 
-public final class SubtractOperator implements Operator{
+public final class SubtractOperator implements Operator {
     @Override
     public final BigDecimal operate(final BigDecimal leftOperand, final BigDecimal rightOperand) {
         return leftOperand.subtract(rightOperand);
