+++ src/test/java/step2/operator/AddOperatorTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.operator;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
@@ -9,30 +9,30 @@ import java.math.BigDecimal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class MultiplyOperatorTest {
+class AddOperatorTest {
 
     private static Operator operator;
 
     @BeforeAll
     static void setUp() {
-        operator = new MultiplyOperator();
+        operator = new AddOperator();
     }
 
-    @DisplayName(value = "두 수에 대한 곱셈 테스트")
+    @DisplayName(value = "두 수에 대한 덧셈 테스트")
     @ParameterizedTest(name = "{index}. 피연산자: [{0},{1}] 결과: [{2}]")
     @CsvSource(value = {
             "0,0,0",
-            "10,20,200",
-            "-10,20,-200",
-            "-10,-20,200",
-            "10000000000,20000000000,200000000000000000000",
-            "-10000000000,20000000000,-200000000000000000000",
-            "-10000000000,-20000000000,200000000000000000000",
-            "10.7,20.8,222.56",
-            "-10.7,20.8,-222.56",
-            "-10.7,-20.8,222.56"
+            "10,20,30",
+            "-10,20,10",
+            "-10,-20,-30",
+            "10000000000,20000000000,30000000000",
+            "-10000000000,20000000000,10000000000",
+            "-10000000000,-20000000000,-30000000000",
+            "10.7,20.8,31.5",
+            "-10.7,20.8,10.1",
+            "-10.7,-20.8,-31.5"
     })
-    void multiplyTest(String leftOperandString, String rightOperandString, String expectedSumString) {
+    void addTest(String leftOperandString, String rightOperandString, String expectedSumString) {
         // given
         BigDecimal leftOperand = new BigDecimal(leftOperandString);
         BigDecimal rightOperand = new BigDecimal(rightOperandString);
