+++ src/test/java/step2/operator/OperatorSelectorTest.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.operator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
